requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
cloudinary==1.38.0
google-cloud-texttospeech==2.15.0
openai==1.12.0
python-dotenv==1.0.1
emoji==2.10.1
pathlib==1.0.1 